Edit private Attributes fine tunning 2
git-svn-id: http://svn.code.sf.net/p/radiusdesk/code/trunk/rd_cake@483 e9fd4974-e3b9-4200-b821-f6b099dd8586

+++ Controller/DevicesController.php
@@ -7,6 +7,11 @@ class DevicesController extends AppController {
     public $components = array('Aa');
     public $uses       = array('Device','User');
     protected $base    = "Access Providers/Controllers/Devices/"; //Required for AP Rights
+    protected  $read_only_attributes = array(
+            'Rd-User-Type', 'Rd-Device-Owner', 'Rd-Account-Disabled', 'User-Profile', 'Expiration',
+            'Rd-Account-Activation-Time', 'Rd-Not-Track-Acct', 'Rd-Not-Track-Auth', 'Rd-Auth-Type', 
+            'Rd-Cap-Type-Data', 'Rd-Cap-Type-Time' ,'Rd-Realm', 'Cleartext-Password'
+        );
 
     //-------- BASIC CRUD -------------------------------
 
@@ -500,16 +505,6 @@ class DevicesController extends AppController {
 
         $items = array();
 
-        $read_only_attributes = array(
-            'Rd-User-Type', 'Rd-Device-Owner', 'Rd-Account-Disabled', 'User-Profile', 'Expiration',
-            'Rd-Account-Activation-Time', 'Rd-Not-Track-Acct', 'Rd-Not-Track-Auth', 'Rd-Auth-Type', 
-            'Rd-Cap-Type-Data', 'Rd-Cap-Type-Data' , 'Rd-Realm', 'Cleartext-Password'
-        );
-
-       // $exclude_attribues = array(
-       //     'Cleartext-Password'
-       // )
-
         //TODO Check if the owner of this user is in the chain of the APs
         if(isset($this->request->query['username'])){
             $username = $this->request->query['username'];
@@ -517,7 +512,7 @@ class DevicesController extends AppController {
             foreach($q_r as $i){
                 $edit_flag      = true;
                 $delete_flag    = true;
-                if(in_array($i['Radcheck']['attribute'],$read_only_attributes)){
+                if(in_array($i['Radcheck']['attribute'],$this->read_only_attributes)){
                     $edit_flag      = false;
                     $delete_flag    = false;
                 }     
@@ -537,7 +532,7 @@ class DevicesController extends AppController {
             foreach($q_r as $i){
                 $edit_flag      = true;
                 $delete_flag    = true;
-                if(in_array($i['Radreply']['attribute'],$read_only_attributes)){
+                if(in_array($i['Radreply']['attribute'],$this->read_only_attributes)){
                     $edit_flag      = false;
                     $delete_flag    = false;
                 }     
@@ -735,7 +730,7 @@ class DevicesController extends AppController {
 
     public function private_attr_delete(){
 
-        $fail_flag = true;
+        $fail_flag = false;
 
         $rc = ClassRegistry::init('Radcheck');
         $rr = ClassRegistry::init('Radreply');
@@ -743,29 +738,45 @@ class DevicesController extends AppController {
         if(isset($this->data['id'])){   //Single item delete
             $type_id            = explode( '_', $this->request->data['id']);
             if(preg_match("/^chk_/",$this->request->data['id'])){
-                $rc->id = $type_id[1];
-                $rc->delete();
+                //Check if it should not be deleted
+                $qr = $rc->findById($type_id[1]);
+                if($qr){
+                    $name = $qr['Radcheck']['attribute'];
+                    if(in_array($name,$this->read_only_attributes)){
+                        $fail_flag = true;
+                    }else{
+                        $rc->id = $type_id[1];
+                        $rc->delete();
+                    }            
+                }
             }
 
             if(preg_match("/^rpl_/",$this->request->data['id'])){   
                 $rr->id = $type_id[1];
                 $rr->delete();
-            }
-            
-            $fail_flag = false;
+            } 
    
         }else{                          //Assume multiple item delete
+            $fail_flag          = false;
             foreach($this->data as $d){
                 $type_id            = explode( '_', $d['id']);
                 if(preg_match("/^chk_/",$d['id'])){
-                    $rc->id = $type_id[1];
-                    $rc->delete();
+                    //Check if it should not be deleted
+                    $qr = $rc->findById($type_id[1]);
+                    if($qr){
+                        $name = $qr['Radcheck']['attribute'];
+                        if(in_array($name,$this->read_only_attributes)){
+                            $fail_flag = true;
+                        }else{
+                            $rc->id = $type_id[1];
+                            $rc->delete();
+                        }            
+                    }
                 }
                 if(preg_match("/^rpl_/",$d['id'])){   
                     $rr->id = $type_id[1];
                     $rr->delete();
-                }          
-                $fail_flag = false;  
+                }            
             }
         }
