+++ Controller/PermanentUsersController.php
@@ -8,6 +8,12 @@ class PermanentUsersController extends AppController {
     public $components = array('Aa');
     protected $base    = "Access Providers/Controllers/PermanentUsers/";
 
+    protected  $read_only_attributes = array(
+            'Rd-User-Type', 'Rd-Device-Owner', 'Rd-Account-Disabled', 'User-Profile', 'Expiration',
+            'Rd-Account-Activation-Time', 'Rd-Not-Track-Acct', 'Rd-Not-Track-Auth', 'Rd-Auth-Type', 
+            'Rd-Cap-Type-Data', 'Rd-Cap-Type-Time' ,'Rd-Realm', 'Cleartext-Password'
+        );
+
     //--- FROM THE OLD ---
     /* json_index json_add json_del json_view json_edit 
         // json_prepaid_list json_tabs json_send_message csv json_change_profile 
@@ -627,12 +633,6 @@ class PermanentUsersController extends AppController {
 
         $items = array();
 
-        $read_only_attributes = array(
-            'Rd-User-Type', 'Rd-Device-Owner', 'Rd-Account-Disabled', 'User-Profile', 'Expiration',
-            'Rd-Account-Activation-Time', 'Rd-Not-Track-Acct', 'Rd-Not-Track-Auth', 'Rd-Auth-Type', 
-            'Rd-Cap-Type-Data', 'Rd-Cap-Type-Time' ,'Rd-Realm', 'Cleartext-Password'
-        );
-
        // $exclude_attribues = array(
        //     'Cleartext-Password'
        // )
@@ -644,7 +644,7 @@ class PermanentUsersController extends AppController {
             foreach($q_r as $i){
                 $edit_flag      = true;
                 $delete_flag    = true;
-                if(in_array($i['Radcheck']['attribute'],$read_only_attributes)){
+                if(in_array($i['Radcheck']['attribute'],$this->read_only_attributes)){
                     $edit_flag      = false;
                     $delete_flag    = false;
                 }     
@@ -664,7 +664,7 @@ class PermanentUsersController extends AppController {
             foreach($q_r as $i){
                 $edit_flag      = true;
                 $delete_flag    = true;
-                if(in_array($i['Radreply']['attribute'],$read_only_attributes)){
+                if(in_array($i['Radreply']['attribute'],$this->read_only_attributes)){
                     $edit_flag      = false;
                     $delete_flag    = false;
                 }     
@@ -869,30 +869,51 @@ class PermanentUsersController extends AppController {
 
         if(isset($this->data['id'])){   //Single item delete
             $type_id            = explode( '_', $this->request->data['id']);
+            $fail_flag          = false;
+
             if(preg_match("/^chk_/",$this->request->data['id'])){
-                $rc->id = $type_id[1];
-                $rc->delete();
+                
+                //Check if it should not be deleted
+                $qr = $rc->findById($type_id[1]);
+                if($qr){
+                    $name = $qr['Radcheck']['attribute'];
+                    if(in_array($name,$this->read_only_attributes)){
+                        $fail_flag = true;
+                    }else{
+                        $rc->id = $type_id[1];
+                        $rc->delete();
+                    }            
+                }
             }
 
             if(preg_match("/^rpl_/",$this->request->data['id'])){   
                 $rr->id = $type_id[1];
                 $rr->delete();
-            }
-            
-            $fail_flag = false;
+            }         
    
-        }else{                          //Assume multiple item delete
+        }else{ 
+            $fail_flag          = false; 
+                        //Assume multiple item delete
             foreach($this->data as $d){
                 $type_id            = explode( '_', $d['id']);
                 if(preg_match("/^chk_/",$d['id'])){
-                    $rc->id = $type_id[1];
-                    $rc->delete();
+
+                    //Check if it should not be deleted
+                    $qr = $rc->findById($type_id[1]);
+                    if($qr){
+                        $name = $qr['Radcheck']['attribute'];
+                        if(in_array($name,$this->read_only_attributes)){
+                            $fail_flag = true;
+                        }else{
+                            $rc->id = $type_id[1];
+                            $rc->delete();
+                        }            
+                    }
                 }
                 if(preg_match("/^rpl_/",$d['id'])){   
                     $rr->id = $type_id[1];
                     $rr->delete();
-                }          
-                $fail_flag = false;  
+                }           
             }
         }
